refactor(apiResponse): 패키지 생성
